DEPI_final/requirements.txt
torch
torchvision
opencv-python
tqdm
scikit-learn
matplotlib
pillow
numpy   